added more input validation + added utils file and separated out some utility functions
