+++ secretsanta/src/components/App/App.css
@@ -229,3 +229,17 @@ button:focus-visible {
 .success-message{
   width: 100%;
 }
+.error-container{
+  height: 50%;
+  align-items: center;
+  justify-content: center;
+
+}
+.error-message{
+  display: block;
+  background-color: rgb(158, 158, 158);
+  width: 100%;
+  padding: 2px;
+  border-radius: 10px;
+  color: #FFFFFF;
+}
